quotable_streamlit/requirements.txt
streamlit
openai
pandas
plotly
crawl4ai
python-dotenv